Split create_api_app() into per-subsystem Flask blueprints
Phase 3 of the hardware_service.py split.  All ~1,500 lines of route
definitions that used to live inline in ``create_api_app()`` move into
matching ``services.<subsystem>.api`` modules; ``create_api_app()`` is
now a thin assembler that builds each blueprint with the
orchestrator-owned runtime state it needs and registers it on the
Flask app.

Per-subsystem blueprint factories
---------------------------------
* services/network/api.py
  - new package (``services.network``) — sibling to displays/gpio/gps/
    zigbee in Phase 2.
  - ``create_blueprint()`` takes no arguments; the 19 NetworkManager /
    nmcli proxy routes (+ /hostname GET/POST = 20 endpoints) are pure
    wrappers around ``app_utils.network`` helpers and need no
    orchestrator state.
* services/zigbee/api.py
  - ``create_blueprint(get_zigpy_controller=...)`` — 6 routes (ports,
    test_port, detect, permit_join POST+DELETE, join_status).  The
    getter callback lets the routes pick up controller restarts
    without re-registration.
  - also fixes a Phase-2-introduced NameError on
    ``/api/zigbee/detect?debug=1``: the detection signatures
    (``_ZIGBEE_USB_SIGNATURES`` / ``_ZIGBEE_BYID_KEYWORDS``) moved into
    ``services.zigbee.detection`` in Phase 2 but the debug branch
    here still referenced the now-deleted module-level names.  The
    blueprint imports them directly from detection.
* services/gps/api.py
  - ``create_blueprint(get_gps_manager=..., get_redis_client=...,
    restart_gps_manager=...)`` — 3 routes (status, trends, configure).
  - ``configure_gps`` now delegates the stop-and-restart dance to
    a single ``restart_gps_manager(enabled)`` callback that lives in
    the orchestrator and wraps with ``_flask_app.app_context()`` —
    the same context the working ``push_screen_to_display`` route
    used.
* services/displays/api.py
  - ``create_blueprint(get_flask_app=...)`` — the single
    ``/api/hardware/display/push`` endpoint used by the web service
    to push DisplayScreens to OLED / LED / VFD hardware.

hardware_service.py shrinks from ~2,000 to ~550 lines.  Only orchestrator
concerns remain: controller-init wrappers, the GPS trend sampler shim
(kept so ``tests/test_gps_trends_archive.py`` keeps working), the
metrics publisher, the health-check loop, and ``main()`` with its
shutdown handler.

No behaviour change at the wire level.  All 31 route paths (including
the bare ``/health``) and response shapes are preserved verbatim — the
move is purely organisational.  Process-split (one Python process per
subsystem) lands in Phase 4.

Verified: ``tests/test_gps_trends_archive.py`` (17 tests) and
``tests/test_sdr_capture_tap_and_auto_gain.py`` (9 tests) — the only
two tests that import ``hardware_service`` — both still pass.
